[#244] Refactor dynamics
